app/src/main/java/com/codesoom/assignment/controllers/GlobalExceptionHandler.java
package com.codesoom.assignment.controllers;

import com.codesoom.assignment.dto.ErrorResponse;
import org.springframework.http.HttpHeaders;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.MethodArgumentNotValidException;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.context.request.WebRequest;
import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;

import java.util.List;
import java.util.stream.Collectors;

@ControllerAdvice
public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {

    @Override
    protected ResponseEntity<Object> handleMethodArgumentNotValid(MethodArgumentNotValidException ex, HttpHeaders headers, HttpStatus status, WebRequest request) {

        List<String> errors = ex.getBindingResult().getFieldErrors()
                .stream()
                .map(e -> String.format("[%s] %s" , e.getField() , e.getDefaultMessage()))
                .collect(Collectors.toList());
        ErrorResponse response = ErrorResponse.builder()
                                                .errors(errors)
                                                .status(HttpStatus.BAD_REQUEST)
                                                .build();

        return new ResponseEntity<>(response, response.getStatus());
    }
}
